+++ src/framework/animation/jacobian_solver.cpp
@@ -44,7 +44,7 @@ void JacobianSolver::set_rotation_axis() {
 
 bool JacobianSolver::solve(const Transform& target) {
 
-    uint32_t num_joints = ik_chain.size() - 1; // we don't need to compute the rotations for the end effector jont
+    uint32_t num_joints = (uint32_t)ik_chain.size() - 1; // we don't need to compute the rotations for the end effector jont
     glm::vec3 goal_pos = target.position;
 
     // init the jacabi matrix (DoF * rotation axis)
